requirements.txt
langchain>=1.0.0
langchain-core>=0.3.0
langchain-community>=0.4.0,<0.5.0
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
redis>=5.0.0
tiktoken>=0.7.0
langsmith>=0.3.0
python-dotenv>=1.0.0
